move locobot_mover_utils and locobot/rotation.py to become common robot utils
