+++ agents/locobot/tests/test_memory.py
@@ -4,7 +4,7 @@ Copyright (c) Facebook, Inc. and its affiliates.
 import unittest
 import numpy as np
 from numpy.testing import assert_allclose
-from droidlet.lowlevel.locobot.locobot_mover_utils import xyz_canonical_coords_to_pyrobot_coords
+from droidlet.lowlevel.robot_mover_utils import xyz_canonical_coords_to_pyrobot_coords
 from droidlet.perception.robot import Detection
 from droidlet.shared_data_structs import RGBDepth
 from droidlet.memory.robot.loco_memory import LocoAgentMemory
