+++ droidlet/interpreter/robot/tasks.py
@@ -11,7 +11,7 @@ from droidlet.memory.memory_nodes import TaskNode
 
 from droidlet.interpreter.robot.objects import DanceMovement
 
-from droidlet.lowlevel.locobot.locobot_mover_utils import (
+from droidlet.lowlevel.robot_mover_utils import (
     get_move_target_for_point,
     ARM_HEIGHT,
     get_camera_angles,
