+++ droidlet/interpreter/robot/tests/fake_agent.py
@@ -16,7 +16,7 @@ from droidlet.dialog.dialogue_manager import DialogueManager
 from droidlet.dialog.map_to_dialogue_object import DialogueObjectMapper
 from droidlet.memory.robot.loco_memory import LocoAgentMemory
 from droidlet.memory.robot.loco_memory_nodes import DetectedObjectNode
-from droidlet.lowlevel.locobot.locobot_mover_utils import (
+from droidlet.lowlevel.robot_mover_utils import (
     get_camera_angles,
     angle_diff,
     CAMERA_HEIGHT,
@@ -25,7 +25,7 @@ from droidlet.lowlevel.locobot.locobot_mover_utils import (
 
 # FXIME!!! everything here should be essentially self-contained
 from agents.locobot.self_perception import SelfPerception
-import droidlet.lowlevel.locobot.rotation as rotation
+import droidlet.lowlevel.rotation as rotation
 from droidlet.perception.robot.tests.utils import get_fake_detection
 from droidlet.shared_data_struct.robot_shared_utils import Pos, RobotPerceptionData
