+++ droidlet/shared_data_structs.py
@@ -1,7 +1,7 @@
 import heapq
 import numpy as np
 import time
-from droidlet.lowlevel.locobot.locobot_mover_utils import xyz_pyrobot_to_canonical_coords
+from droidlet.lowlevel.robot_mover_utils import xyz_pyrobot_to_canonical_coords
 
 
 class Time:
